feat: Use pinned top app bar
The top app bar now uses a pinned scroll behavior.
This allows the app bar to remain visible at the top of the screen while scrolling content.
Removed the drawerState parameter from the NavTopAppBar composable.
Added scrollBehavior to the NavTopAppBar composable.
Added scrollBehavior and modifier.nestedScroll to the following screens:
- MainSettingsScreen
- TtsLogScreen
- SpeechRuleManagerScreen
- BasicForwarderScreen
- PluginManagerScreen
- ReplaceRuleManagerScreen
- ListManagerScreen
- ForwarderTopAppBar